+++ extension/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the Tabstronaut extension will be documented in this file
 ## [1.3.5]
 
 - Avoid opening duplicate tabs by switching to an existing tab if it is already open in another editor group.
+- Added button to toggle between collapsing and expanding all Tab Groups.
+- Fixed expand all command when a group filter is active.
+- Default to "Expand all Tab Groups" when all groups start collapsed.
 
 ## [1.3.3-1.3.4]